Исправить тесты в соотвествии с заданием
